Update InteractionController.gd
Added the ability for the player to pick/drop items by pressing g. There is a bug that if you move the item quickly while holding it you won't be able to drop it. The reasoning is that in line 17 none of the code inside its if statement will run unless the raycast3d is on an the object, so the only way to drop the object is by having the raycast3d on the object and then click g.